create folder RegistrationPage and render layout
