+++ src/App/pages/StudentPages/RegistrationPage/Fields/formFields.js
@@ -1,0 +1,24 @@
+
+const defaultValues = {
+	phoneNumber: "",
+	address: "",
+	narrow: "",
+	business: "",
+	dream: "",
+};
+export const schoolSupportDefaultValues = {
+	...defaultValues,
+	upload: "",
+};
+export const selfFindingDefaultValues = {
+	phoneNumber: "",
+	address: "",
+	narrow: "",
+	business: "",
+	dream: "",
+	taxCode: "",
+	phoneNumberCompany: "",
+	emailEnterprise: "",
+	addressCompany: "",
+	position: "",
+};
